♻️ Split code for fullscreen cropper
