+++ src/routes/(app)/(sidepanel)/o/[observation]/classify/Subject.svelte
@@ -117,4 +117,10 @@
 	:global(picture) {
 		height: 100%;
 	}
+
+	img {
+		height: 100%;
+		width: 100%;
+		object-fit: contain;
+	}
 </style>
