+++ src/routes/(app)/(sidepanel)/o/[observation]/crop/[image]/+page.svelte
@@ -1,5 +1,121 @@
+<script lang="ts" module>
+	// import type {PageData} from './'
+	import * as idb from '$lib/idb.svelte.js';
+
+	export interface SelectedBox {
+		imageId: string | null;
+		manual: boolean;
+	}
+
+	export const zoom = $state({ ...INITIAL_ZOOM_STATE });
+
+	export const zoomSpeed = () => zoom.scale * 0.1;
+
+	const { sortedFileIds } = $derived<PageData>(page.data);
+	const fileId = $derived(page.params.image);
+
+	const images = $derived.by(() => {
+		if (!fileId) return [];
+
+		return imagesOfImageFile(
+			fileId,
+			idb.tables.Image.state.filter((img) => img.sessionId === uiState.currentSessionId)
+		);
+	});
+
+
+	export function currentImages() {
+		return images;
+	}
+
+	let focusedImageId = $derived.by(() => {
+		if (!page.params.observation) return;
+
+		const observation = idb.tables.Observation.getFromState(page.params.observation);
+		return observation?.images.find((i) => imageIdToFileId(i) === fileId);
+	});
+
+	export function focusedImage() {
+		if (!focusedImageId) return undefined;
+		return idb.tables.Image.getFromState(focusedImageId);
+	}
+
+	export function setFocusedImage(imageId: string | null) {
+		focusedImageId = imageId ?? undefined;
+	}
+
+	export function boundingBoxes() {
+		return Object.fromEntries(
+			images
+				.map(({ id, metadata }) => [id, metadata[uiState.cropMetadataId]?.value])
+				.filter(([, box]) => box !== undefined)
+				.map(([id, box]) => [id, assertIs('boundingbox', box)])
+		);
+	}
+
+	export function hasCrop(imageFileId: string | undefined) {
+		if (!imageFileId) return false;
+
+		return imagesOfImageFile(imageFileId).every(
+			(image) => uiState.cropMetadataId in image.metadata
+		);
+	}
+
+	export function hasConfirmedCrop(imageFileId: string | undefined) {
+		if (!imageFileId) return false;
+
+		return imagesOfImageFile(imageFileId).every(imageHasConfirmedCrop);
+	}
+
+	export function imageHasConfirmedCrop(image: DB.Image) {
+		const value = uiState.cropMetadataValueOf(image);
+		return value?.confirmed;
+	}
+
+	export async function goToFile(fileId: string | undefined) {
+		if (!fileId) return;
+		await goto('/(app)/(sidepanel)/o/[observation]/crop/[image]', {
+			// Changing files might mean changing observations, so we can't keep the same
+			observation: '_',
+			image: fileId,
+		});
+	}
+
+	const prevFileId = $derived.by(() => {
+		const idx = sortedFileIds.indexOf(fileId) - 1;
+		if (idx < 0) return undefined;
+		return sortedFileIds.at(idx);
+	});
+	const nextFileId = $derived.by(() => {
+		const idx = sortedFileIds.indexOf(fileId) + 1;
+		if (idx >= sortedFileIds.length) return undefined;
+		return sortedFileIds.at(idx);
+	});
+
+	const nextUnconfirmedImageId = $derived.by(() => {
+		const forward = sortedFileIds
+			.slice(sortedFileIds.indexOf(fileId) + 1)
+			.filter((fileId) => !hasConfirmedCrop(fileId))
+			.at(0);
+
+		if (forward) return forward;
+
+		// Loop around.
+		return sortedFileIds.find((fileId) => !hasConfirmedCrop(fileId));
+	});
+
+	export async function goToNextUnconfirmedFile() {
+		if (!nextUnconfirmedImageId) return false;
+		await goToFile(nextUnconfirmedImageId);
+	}
+
+	export function adjacentFileIds() {
+		return { prevFileId, nextFileId };
+	}
+</script>
+
 <script lang="ts">
-	import type { CenteredBoundingBox, Rect } from '$lib/BoundingBoxes.svelte.js';
+	import type { CenteredBoundingBox } from '$lib/BoundingBoxes.svelte.js';
 	import type * as DB from '$lib/database.js';
 	import type { RuntimeValue } from '$lib/schemas/metadata';
 
@@ -17,6 +133,7 @@
 	import IconRevert from '~icons/ri/reset-left-fill';
 	import IconToolDragCrop from '~icons/ri/shape-2-line';
 	import IconNeuralNet from '~icons/ri/sparkling-line';
+	import { page } from '$app/state';
 	import {
 		boundingBoxIsNonZero,
 		coordsAreEqual,
@@ -32,714 +149,40 @@
 	import DraggableBoundingBox from '$lib/DraggableBoundingBox.svelte';
 	import { INITIAL_ZOOM_STATE } from '$lib/DraggableBoundingBox.svelte.js';
 	import { percent } from '$lib/i18n';
-	import * as idb from '$lib/idb.svelte.js';
-	import {
-		deleteImageFile,
-		imageIdToFileId,
-		imagesOfImageFile,
-		imageId as makeImageId,
-		parseImageId,
-	} from '$lib/images.js';
-	import { defineKeyboardShortcuts } from '$lib/keyboard.svelte.js';
-	import LoadingSpinner from '$lib/LoadingSpinner.svelte';
-	import {
-		assertIs,
-		deleteMetadataValue,
-		hasRuntimeType,
-		storeMetadataValue,
-	} from '$lib/metadata/index.js';
+	import { imageIdToFileId, imagesOfImageFile } from '$lib/images.js';
+	import { assertIs } from '$lib/metadata/index.js';
 	import MobileWIPOverlay from '$lib/MobileWIPOverlay.svelte';
-	import { ensureNoEmptyObservations } from '$lib/observations.js';
-	import OverflowableText from '$lib/OverflowableText.svelte';
 	import { goto } from '$lib/paths.js';
-	import SentenceJoin from '$lib/SentenceJoin.svelte';
 	import { seo } from '$lib/seo.svelte';
-	import { getSettings, toggleSetting } from '$lib/settings.svelte';
 	import { uiState } from '$lib/state.svelte';
-	import { toasts } from '$lib/toasts.svelte';
-	import Tooltip from '$lib/Tooltip.svelte';
-	import { tooltip } from '$lib/tooltips';
-	import { undo } from '$lib/undo.svelte';
-	import {
-		clamp,
-		fromEntries,
-		mapValues,
-		nonnull,
-		pick,
-		range,
-		sign,
-		throwError,
-	} from '$lib/utils';
+	import { mapValues } from '$lib/utils';
 
 	import TopbarExtras from '../../TopbarExtras.svelte';
+	import { changeAllConfirmedStatuses, setupUndoActions } from './actions.svelte.js';
+	import Boxes from './Boxes.svelte';
+	import CropSurface from './CropSurface.svelte';
+	import { setupKeyboardShortcuts } from './keyboard.svelte.ts';
+	import Toolbar from './Toolbar.svelte';
 
 	const { data, params } = $props();
-	const { sortedFileIds } = $derived(data);
 
-	const fileId = $derived(params.image);
-
-	const images = $derived(
-		imagesOfImageFile(
-			fileId,
-			idb.tables.Image.state.filter((img) => img.sessionId === uiState.currentSessionId)
-		)
-	);
 	const firstImage = $derived(images.at(0));
-
-	seo({ title: `Recadrer ${firstImage?.filename ?? '...'}` });
 
 	// Controls visibility of the checkmark little centered overlay
 	let showConfirmedOverlay = $state(async () => {});
-
-	let activeToolName = $state<(typeof tools)[number]['name']>('Glisser-recadrer');
-
-	interface Tool {
-		name: string;
-		help: string;
-		icon: import('svelte').Component;
-		shortcut: string;
-		transformable: boolean;
-		createMode: 'clickanddrag' | '2point' | '4point' | 'off';
-		movable: boolean;
-		cursor?: string;
-	}
-
-	const tools = [
-		{
-			name: 'Glisser-recadrer',
-			help: 'Cliquer et glisser pour créer une boîte de recadrage',
-			icon: IconToolDragCrop,
-			shortcut: 'r',
-			transformable: true,
-			createMode: 'clickanddrag',
-			movable: true,
-			cursor: 'crosshair',
-		},
-		{
-			name: '2 points',
-			help: 'Cliquer sur les 2 coins pour créer une boîte de recadrage',
-			icon: IconTwoPointCrop,
-			shortcut: 'z',
-			transformable: false,
-			createMode: '2point',
-			movable: false,
-			cursor: 'crosshair',
-		},
-		{
-			name: '4 points',
-			help: 'Cliquer sur les 4 extrémités pour créer une boîte de recadrage',
-			icon: IconFourPointCrop,
-			shortcut: 'Shift+z',
-			transformable: false,
-			createMode: '4point',
-			movable: false,
-			cursor: 'crosshair',
-		},
-		{
-			name: 'Déplacer',
-			help: 'Cliquer et glisser pour déplacer la boîte de recadrage',
-			icon: IconToolMove,
-			shortcut: 'v',
-			transformable: false,
-			createMode: 'off',
-			movable: true,
-			cursor: 'pointer',
-		},
-		{
-			name: 'Main',
-			help: "Cliquer et glisser pour se déplacer dans l'image",
-			icon: IconToolHand,
-			shortcut: 'h',
-			transformable: false,
-			createMode: 'off',
-			movable: false,
-			cursor: 'grab',
-		},
-	] as const satisfies Tool[];
-
-	const creationTools = $derived(tools.filter(({ createMode }) => createMode !== 'off'));
-
-	let activeTool = $derived(tools.find(({ name }) => name === activeToolName) || tools[0]);
-
-	let focusedImageId = $derived(
-		params.observation
-			? idb.tables.Observation.getFromState(params.observation)?.images.find(
-					(i) => imageIdToFileId(i) === params.image
-				)
-			: ''
-	);
-
-	const boundingBoxes = $derived<Record<string, RuntimeValue<'boundingbox'>>>(
-		Object.fromEntries(
-			images
-				.map(({ id, metadata }) => [id, metadata[uiState.cropMetadataId]?.value])
-				.filter(([, box]) => box !== undefined)
-				.map(([id, box]) => [id, assertIs('boundingbox', box)])
-		)
-	);
-
-	const imageSrc = $derived(uiState.getPreviewURL(fileId));
-	const prevFileId = $derived.by(() => {
-		const idx = sortedFileIds.indexOf(fileId) - 1;
-		if (idx < 0) return undefined;
-		return sortedFileIds.at(idx);
-	});
-	const nextFileId = $derived.by(() => {
-		const idx = sortedFileIds.indexOf(fileId) + 1;
-		if (idx >= sortedFileIds.length) return undefined;
-		return sortedFileIds.at(idx);
-	});
-	const nextUnconfirmedImageId = $derived.by(() => {
-		const forward = sortedFileIds
-			.slice(sortedFileIds.indexOf(fileId) + 1)
-			.filter((fileId) => !hasConfirmedCrop(fileId))
-			.at(0);
-
-		if (forward) return forward;
-
-		// Loop around.
-		return sortedFileIds.find((fileId) => !hasConfirmedCrop(fileId));
-	});
-
-	const initialCrops = $derived<
-		Record<string, undefined | { value: RuntimeValue<'boundingbox'>; confidence: number }>
-	>(
-		Object.fromEntries(
-			images.map((image) => {
-				if (!image.metadata[uiState.cropMetadataId]) {
-					return [image.id, undefined];
-				}
-
-				const { alternatives, value, confidence, manuallyModified } =
-					image.metadata[uiState.cropMetadataId];
-
-				// On subsequent crops, the user's crop will be the main value and the neural network's crop will be in the alternatives.
-				if (alternatives && Object.entries(alternatives).length > 0) {
-					const [[stringValue, confidence]] = Object.entries(alternatives);
-					return [
-						image.id,
-						{
-							value: JSON.parse(stringValue),
-							confidence,
-						},
-					];
-				}
-
-				// If the main value is manuallyModified, and we have no alternatives,
-				// it means that the box never had a neural network-inferred value (or it was lost somehow)
-				if (manuallyModified) {
-					return [image.id, undefined];
-				}
-
-				// If this is the first time the user is re-cropping the box, this value will be the main values.
-				return [image.id, { value, confidence }];
-			})
-		)
-	);
-
-	async function goToFile(fileId: string | undefined) {
-		if (!fileId) return;
-		await goto('/(app)/(sidepanel)/o/[observation]/crop/[image]', {
-			// Changing files might mean changing observations, so we can't keep the same
-			observation: '_',
-			image: fileId,
-		});
-	}
-
-	function hasCrop(imageFileId: string) {
-		return imagesOfImageFile(imageFileId).every(
-			(image) => uiState.cropMetadataId in image.metadata
-		);
-	}
-
-	function hasConfirmedCrop(imageFileId: string | undefined) {
-		if (!imageFileId) return false;
-		return imagesOfImageFile(imageFileId).every(imageHasConfirmedCrop);
-	}
-
-	function imageHasConfirmedCrop(image: DB.Image) {
-		const value = uiState.cropMetadataValueOf(image);
-		return value?.confirmed;
-	}
-
-	async function changeCropConfirmedStatus(image: DB.Image, confirmed: boolean) {
-		if (!uiState.currentSessionId) return;
-		if (!uiState.cropMetadataId) return;
-
-		const value = uiState.cropMetadataValueOf(image) ?? {
-			confidence: 1,
-			manuallyModified: true,
-			value: FULL_IMAGE_CROPBOX,
-			confidences: {},
-			confirmed: true,
-		};
-
-		await storeMetadataValue({
-			db: idb.databaseHandle(),
-			sessionId: uiState.currentSessionId,
-			metadataId: uiState.cropMetadataId,
-			subjectId: image.id,
-			type: 'boundingbox',
-			...value,
-			confirmed,
-		});
-	}
-
-	async function changeAllConfirmedStatuses(confirmed: boolean) {
-		for (const image of images) {
-			await changeCropConfirmedStatus(image, confirmed);
-		}
-	}
-
-	async function revertToInferredCrop(imageId: string) {
-		const initialCrop = initialCrops[imageId];
-		// On subsequent crops, the user's crop will be the main value and the neural network's crop will be in the alternatives.
-		if (!initialCrop) {
-			toasts.error(
-				`L'image ${imageId} n'a pas de recadrage alternatif, impossible de revenir au recadrage d'origine`
-			);
-			return;
-		}
-
-		if (!uiState.currentSessionId) return;
-
-		await storeMetadataValue({
-			db: idb.databaseHandle(),
-			sessionId: uiState.currentSessionId,
-			subjectId: imageId,
-			metadataId: uiState.cropMetadataId,
-			type: 'boundingbox',
-			manuallyModified: false,
-			...initialCrop,
-		});
-
-		if (uiState.classificationMetadataId) {
-			await deleteMetadataValue({
-				db: idb.databaseHandle(),
-				sessionId: uiState.currentSessionId,
-				metadataId: uiState.classificationMetadataId,
-				subjectId: imageId,
-			});
-		}
-	}
-
-	const revertableCrops = $derived(
-		Object.fromEntries(
-			images.map((image) => {
-				const initial = initialCrops[image.id]?.value;
-				const current = image.metadata[uiState.cropMetadataId]?.value;
-				return [
-					image.id,
-					initial &&
-						hasRuntimeType('boundingbox', current) &&
-						!coordsAreEqual(initial, current),
-				];
-			})
-		)
-	);
-
-	// We can revertAll() if either we have no initial crops at all, or at least some of them are revertable
-	const canRevertAll = $derived(
-		Object.values(initialCrops).every((c) => !c) || Object.values(revertableCrops).some(Boolean)
-	);
-
-	async function revertAll() {
-		// Either we have no initial crop: reverting means removing all boxes
-		if (Object.values(initialCrops).every((c) => !c)) {
-			for (const { id } of images) {
-				await deleteBoundingBox(id);
-			}
-		} else {
-			// Or we have at least one: revert all boxes to their initial positions
-			for (const { id } of images) {
-				await revertToInferredCrop(id);
-			}
-		}
-	}
-
-	undo.on('crop/box/create', async ({ imageId }) => {
-		if (imageIdToFileId(imageId) !== fileId) return;
-		await deleteBoundingBox(imageId, { skipUndo: true });
-	});
-
-	undo.on('crop/box/edit', async ({ imageId, before }) => {
-		if (imageIdToFileId(imageId) !== fileId) return;
-		await onCropChange(imageId, toTopLeftCoords(before), false, false);
-	});
-
-	undo.on('crop/box/delete', async ({ imageId, box }) => {
-		if (imageIdToFileId(imageId) !== fileId) return;
-		await onCropChange(null, toTopLeftCoords(box), false, false);
-	});
-
-	/**
-	 * @param options
-	 * @param options.skipUndo whether to skip pushing this operation to the undo stack
-	 */
-	async function deleteBoundingBox(
-		imageId: string,
-		{ skipUndo = false }: { skipUndo?: boolean } = {}
-	) {
-		if (!skipUndo) {
-			undo.push('crop/box/delete', {
-				imageId,
-				box: boundingBoxes[imageId],
-			});
-		}
-
-		if (images.length === 1) {
-			await deleteMetadataValue({
-				db: idb.databaseHandle(),
-				sessionId: uiState.currentSession?.id,
-				metadataId: uiState.cropMetadataId,
-				subjectId: imageId,
-			});
-		} else {
-			await idb.tables.Image.remove(imageId);
-		}
-
-		await ensureNoEmptyObservations();
-	}
-
-	/**
-	 * @param  imageId ID of the image we're confirming a new crop for. Null if we're creating a new cropbox.
-	 * @param  newBoundingBox
-	 * @param  flashConfirmedOverlay flash the confirmed overlay when appropriate
-	 * @param  pushToUndoStack whether to push this change to the undo stack
-	 * @returns  the ID of the image we just modified/created
-	 */
-	async function onCropChange(
-		imageId: string | null,
-		newBoundingBox: Rect | undefined,
-		flashConfirmedOverlay = true,
-		pushToUndoStack = true
-	): Promise<string | null> {
-		if (!uiState.currentSessionId) return null;
-
-		const image = imageId ? images.find((img) => img.id === imageId) : undefined;
-		if (imageId && !image) {
-			toasts.error(`Impossible de trouver l'image ${imageId} pour la confirmer`);
-			return null;
-		}
-
-		if (!newBoundingBox) {
-			// No bounding box, just mark the image as confirmed and move on
-			if (image) {
-				await changeCropConfirmedStatus(image, true);
-			}
-
-			return imageId;
-		}
-
-		// Flash if
-		const willFlashConfirmedOverlay =
-			// the caller asked for it,
-			flashConfirmedOverlay &&
-			// and this is the last image before the file is considered confirmed
-			images.filter(imageHasConfirmedCrop).length === images.length - 1;
-
-		const willAutoskip =
-			// The user has auto-skip enabled
-			getSettings().cropAutoNext &&
-			// We aren't deleting the bounding box
-			boundingBoxIsNonZero(newBoundingBox) &&
-			// in click-and-drag creation mode, auto-skip occurs only when the bounding box did not exist before (since the first creation of the box is much less likely to be the final one)
-			(activeTool.createMode !== 'clickanddrag' || !hasCrop(fileId));
-
-		if (uiState.classificationMetadataId) {
-			const species = image?.metadata[uiState.classificationMetadataId];
-			if (species && !species.manuallyModified) {
-				// Species confidence was inferred, we need to remove it so we can infer it again, since it's inferred on the _cropped_ image
-				await deleteMetadataValue({
-					db: idb.databaseHandle(),
-					sessionId: uiState.currentSessionId,
-					metadataId: uiState.classificationMetadataId,
-					subjectId: image.id,
-				});
-			}
-		}
-
-		let newImageId = '';
-
-		if (imageId) {
-			if (pushToUndoStack) {
-				undo.push('crop/box/edit', {
-					imageId,
-					before: boundingBoxes[imageId],
-					after: toCenteredCoords(newBoundingBox),
-				});
-			}
-
-			// We're modifying an existing cropbox
-			await storeMetadataValue({
-				db: idb.databaseHandle(),
-				sessionId: uiState.currentSessionId,
-				metadataId: uiState.cropMetadataId,
-				subjectId: imageId,
-				type: 'boundingbox',
-				value: toCenteredCoords(newBoundingBox),
-				confidence: 1,
-				// Put the neural-network-inferred (initial) value in the alternatives as a backup
-				confidences: initialCrops[imageId] ? [initialCrops[imageId]] : [],
-				manuallyModified: true,
-			});
-		} else if (
-			images.length === 1 &&
-			firstImage &&
-			!firstImage.metadata[uiState.cropMetadataId]
-		) {
-			// We're creating a new cropbox, but it is the first one (and we already have an image, it just doesn't have a cropbox)
-			newImageId = firstImage.id;
-
-			if (pushToUndoStack) {
-				undo.push('crop/box/create', {
-					imageId: newImageId,
-					box: toCenteredCoords(newBoundingBox),
-				});
-			}
-
-			await storeMetadataValue({
-				db: idb.databaseHandle(),
-				sessionId: uiState.currentSessionId,
-				metadataId: uiState.cropMetadataId,
-				subjectId: newImageId,
-				type: 'boundingbox',
-				value: toCenteredCoords(newBoundingBox),
-				confidence: 1,
-				manuallyModified: true,
-				confidences: [],
-			});
-		} else {
-			// We're creating a >1st cropbox
-			newImageId = makeImageId(
-				fileId,
-				Math.max(...images.map(({ id }) => parseImageId(id).subindex).filter(nonnull)) + 1
-			);
-
-			if (pushToUndoStack) {
-				undo.push('crop/box/create', {
-					imageId: newImageId,
-					box: toCenteredCoords(newBoundingBox),
-				});
-			}
-
-			await idb.tables.Image.set({
-				id: newImageId,
-				sessionId: uiState.currentSessionId,
-				filename: firstImage?.filename ?? '',
-				addedAt: dates.formatISO(firstImage?.addedAt ?? new Date()),
-				contentType: firstImage?.contentType ?? '',
-				dimensions: $state.snapshot(firstImage?.dimensions) ?? { width: 0, height: 0 },
-				fileId,
-				metadata: {
-					[uiState.cropMetadataId]: {
-						value: JSON.stringify(toCenteredCoords(newBoundingBox)),
-						confidence: 1,
-						manuallyModified: true,
-						confirmed: true,
-						alternatives: [],
-						confidences: {},
-					},
-				},
-			});
-		}
-
-		showBoxesListHint = false;
-
-		await changeAllConfirmedStatuses(true);
-
-		// Select cropbox
-		if (!selectedBox.manual) {
-			selectedBox.imageId = newImageId || imageId;
-		}
-
-		if (willFlashConfirmedOverlay) {
-			await showConfirmedOverlay();
-		}
-
-		if (willAutoskip) {
-			if (nextUnconfirmedImageId) {
-				await goToFile(nextUnconfirmedImageId);
-			} else {
-				await goto('/(app)/(sidepanel)/classify');
-			}
-		}
-
-		return newImageId;
-	}
-
-	async function deleteImageFileAndGotoNext() {
-		const nextFileIdBeforeDelete = $state.snapshot(nextFileId);
-		await deleteImageFile(fileId);
-
-		if (nextFileIdBeforeDelete) {
-			await goToFile(nextFileIdBeforeDelete);
-		} else {
-			// If nextFileId (and not nextFileIdBeforeDelete) is undefined,
-			// it means we just deleted the last image; so we go back to the import tab
-			await goto('/(app)/(sidepanel)/import');
-		}
-	}
-
-	/**
-	 * @returns  pixel dimensions of the box
-	 */
-	function roundedPixelDimensions(box: CenteredBoundingBox): [number, number] {
-		if (!firstImage) return [0, 0];
-		const scaler = coordsScaler({
-			x: firstImage.dimensions.width,
-			y: firstImage.dimensions.height,
-		});
-
-		const { w, h } = mapValues(scaler(box), Math.round);
-		return [w, h];
-	}
-
-	$effect(() => {
-		uiState.imageOpenedInCropper = fileId;
-	});
-
-	interface SelectedBox {
-		imageId: string | null;
-		manual: boolean;
-	}
 
 	let selectedBox = $state<SelectedBox>({
 		imageId: null,
 		manual: false,
 	});
 
-	defineKeyboardShortcuts('cropping', {
-		'$mod+Delete': {
-			help: 'Supprimer l’image',
-			do: deleteImageFileAndGotoNext,
-		},
-		a: {
-			help: 'Activer/désactiver la continuation automatique',
-			do: async () => toggleSetting('cropAutoNext'),
-		},
-		Delete: {
-			help: 'Supprimer la boîte sélectionnée',
-			when: () => Boolean(selectedBox.imageId),
-			async do() {
-				if (!selectedBox.imageId) return;
-				await deleteBoundingBox(selectedBox.imageId);
+	seo({ title: `Recadrer ${firstImage?.filename ?? '...'}` });
 
-				if (selectedBox.manual) {
-					selectedBox.imageId = null;
-				} else {
-					// Select previous box in list
-					selectedBox.imageId =
-						images.toReversed().find((image) => image.id in boundingBoxes)?.id ?? null;
-				}
-			},
-		},
-		f: {
-			help: 'Cacher les boîtes non sélectionnées',
-			when: () => Boolean(selectedBox),
-			do() {
-				if (!selectedBox) return;
-				if (selectedBox.imageId === focusedImageId) {
-					focusedImageId = '';
-				} else if (selectedBox.imageId) {
-					focusedImageId = selectedBox.imageId;
-				}
-			},
-		},
-		u: {
-			help: "Revenir au recadrage d'origine",
-			when: () => Boolean(revertableCrops[fileId]),
-			do: () => {
-				if (!revertableCrops[fileId]) return;
-				revertToInferredCrop(fileId);
-			},
-		},
-		'$mod+u': {
-			help: "Revenir au recadrage d'origine pour toutes les boîtes",
-			when: () => Object.keys(boundingBoxes).length > 0,
-			do: revertAll,
-		},
+	setupUndoActions({ selectedBox, showConfirmedOverlay });
+	setupKeyboardShortcuts({ selectedBox });
 
-		'+': {
-			help: 'Zoomer',
-			do: () => {
-				zoom.scale = clamp(1, zoom.scale + 4 * zoomSpeed, 10);
-			},
-		},
-		'-': {
-			help: 'Dézoomer',
-			do: () => {
-				zoom.scale = clamp(1, zoom.scale - 4 * zoomSpeed, 10);
-			},
-		},
-		Digit0: {
-			help: 'Réinitialiser le zoom',
-			do: () => {
-				zoom.origin = { x: 0, y: 0 };
-				zoom.scale = 1;
-			},
-		},
-		...fromEntries(
-			tools.map((tool) => [
-				tool.shortcut,
-				{
-					help: `Choisir l'outil ${tool.name}`,
-					do: () => {
-						activeToolName = tool.name;
-					},
-				},
-			])
-		),
-		',': {
-			help: 'Sélectionner la boîte précédente',
-			do: () => {
-				const imageIds = Object.keys(boundingBoxes);
-				const currentIndex = imageIds.indexOf(selectedBox.imageId ?? '');
-				const prevIndex = (currentIndex - 1 + imageIds.length) % imageIds.length;
-				selectedBox.imageId = imageIds[prevIndex];
-			},
-		},
-		';': {
-			help: 'Sélectionner la boîte suivante',
-			do: () => {
-				const imageIds = Object.keys(boundingBoxes);
-				const currentIndex = imageIds.indexOf(selectedBox.imageId ?? '');
-				const nextIndex = (currentIndex + 1) % imageIds.length;
-				selectedBox.imageId = imageIds[nextIndex];
-			},
-		},
-		...fromEntries(
-			range(1, 10).map((i) => [
-				`Digit${i}`,
-				{
-					help: `Sélectionner la boîte #${i}`,
-					when: () => Object.keys(boundingBoxes).length >= i,
-					do: () => {
-						const imageId = Object.keys(boundingBoxes)[i - 1];
-						if (selectedBox.imageId === imageId) {
-							selectedBox.imageId = null;
-							selectedBox.manual = false;
-						} else {
-							selectedBox.imageId = imageId;
-							selectedBox.manual = true;
-						}
-					},
-				},
-			])
-		),
-		'x b': {
-			help: 'Entrer des coordonnées de boîte',
-			async do() {
-				const [x, y, w, h] = (
-					prompt('Coords? (x space y space w space h, ∈ [0, 1])') ??
-					throwError('Prompt annulé')
-				)
-					.split(' ')
-					.map(Number.parseFloat);
-
-				await onCropChange(null, toTopLeftCoords({ x, y, w, h }));
-			},
-		},
+	$effect(() => {
+		uiState.imageOpenedInCropper = fileId;
 	});
 
 	// Scroll to selected box
@@ -752,7 +195,6 @@
 		});
 	});
 
-	let imageElement = $state();
 	let imageIsLoading = $state(true);
 
 	watch(
@@ -760,17 +202,16 @@
 		(newFileId, oldFileId) => {
 			if (oldFileId) uiState.cropperZoomStates.set(oldFileId, $state.snapshot(zoom));
 
-			zoom = $state.snapshot(uiState.cropperZoomStates.get(newFileId)) ?? INITIAL_ZOOM_STATE;
+			const newZoom =
+				$state.snapshot(uiState.cropperZoomStates.get(newFileId ?? '')) ??
+				INITIAL_ZOOM_STATE;
+
+			// We cant reassign to zoom since its an exported $state (so cannot declare it with `let` )
+			for (const prop in newZoom) {
+				zoom[prop] = newZoom[prop];
+			}
 		}
 	);
-
-	let zoom = $state({ ...INITIAL_ZOOM_STATE });
-
-	let showBoxesListHint = $derived(
-		images.every(({ metadata }) => !metadata[uiState.cropMetadataId]?.manuallyModified)
-	);
-
-	const zoomSpeed = $derived(zoom.scale * 0.1);
 </script>
 
 <MobileWIPOverlay
@@ -812,255 +253,22 @@
 			await goToFile(nextFileId);
 		},
 		async nextUnconfirmed() {
-			(await nextUnconfirmedImageId)
-				? goToFile(nextUnconfirmedImageId)
-				: goto('/(app)/(sidepanel)/classify');
+			(await goToNextUnconfirmedFile()) || goto('/(app)/(sidepanel)/classify');
 		},
 	}}
 />
 
 <div class="layout">
-	<!-- svelte-ignore a11y_no_noninteractive_element_interactions -->
-	<main
-		class="crop-surface"
-		onmousedown={async (e) => {
-			// Pan on mousewhell button hold or hand tool
-			if (activeTool.name !== 'Main' && e.button !== 1) return;
+	<div class="crop-surface">
+		<CropSurface bind:imageIsLoading {showConfirmedOverlay} {selectedBox} />
+	</div>
 
-			// Hide autoscroll indicator on Firefox
-			e.preventDefault();
-
-			zoom.panning = true;
-			zoom.panStart = {
-				x: e.clientX,
-				y: e.clientY,
-				zoomOrigin: $state.snapshot(zoom.origin),
-			};
-		}}
-		onmouseup={async ({ button }) => {
-			// Pan on mousewheel button release or hand tool
-			if (activeTool.name !== 'Main' && button !== 1) return;
-			zoom.panning = false;
-		}}
-		onmousemove={async ({ clientX, clientY }) => {
-			if (!zoom.panning) return;
-
-			zoom.origin.x = zoom.panStart.zoomOrigin.x + (clientX - zoom.panStart.x);
-			zoom.origin.y = zoom.panStart.zoomOrigin.y + (clientY - zoom.panStart.y);
-		}}
-		onwheel={async (e) => {
-			e.preventDefault();
-			// Most logic is thanks to https://stackoverflow.com/a/70251437
-			let imageBounds = imageElement.getBoundingClientRect();
-			let x = (e.clientX - imageBounds.x) / zoom.scale;
-			let y = (e.clientY - imageBounds.y) / zoom.scale;
-
-			zoom.scale = clamp(1, zoom.scale - sign(e.deltaY) * 2 * zoomSpeed, 10);
-
-			if (zoom.scale > 1) {
-				zoom.origin.x += sign(e.deltaY) * zoomSpeed * (x * 2 - imageElement.offsetWidth);
-				zoom.origin.y += sign(e.deltaY) * zoomSpeed * (y * 2 - imageElement.offsetHeight);
-			} else {
-				zoom.origin = { x: 0, y: 0 };
-			}
-		}}
-	>
-		<div class="behind-image">
-			{#if imageIsLoading}
-				<LoadingSpinner --size="2em" />
-				<p class="loading">Chargement de l'image…</p>
-			{:else}
-				<p class="coucou" aria-hidden="true">
-					coucou toi :)
-					<br />
-					passes une bonne journée ! 💖💖
-				</p>
-			{/if}
-		</div>
-		<img
-			pw-testid="crop-subject-image"
-			src={imageSrc}
-			alt=""
-			bind:this={imageElement}
-			style:scale={zoom.scale}
-			style:translate="{zoom.origin.x}px {zoom.origin.y}px"
-			onload={() => {
-				imageIsLoading = false;
-			}}
-		/>
-		{#if imageElement}
-			<DraggableBoundingBox
-				{...activeTool}
-				{imageElement}
-				{zoom}
-				imageFileID={fileId}
-				boundingBoxes={mapValues(
-					focusedImageId ? pick(boundingBoxes, focusedImageId) : boundingBoxes,
-					toTopLeftCoords
-				)}
-				disabled={zoom.panning}
-				cursor={zoom.panning
-					? activeTool.name === 'Main'
-						? 'grabbing'
-						: 'move'
-					: activeTool.cursor}
-				onchange={(imageId, box) => onCropChange(imageId, box)}
-				oncreate={(box) => onCropChange(null, box)}
-			/>
-		{/if}
-	</main>
 	<aside class="toolbar">
-		{#each tools as tool (tool.name)}
-			<button
-				aria-label="Choisir l'outil {tool.name}"
-				class:active={tool.name === activeToolName}
-				use:tooltip={{
-					text: `${tool.name}: ${tool.help}`,
-					keyboard: tool.shortcut,
-					placement: 'right',
-				}}
-				onclick={() => {
-					activeToolName = tool.name;
-				}}
-			>
-				<tool.icon />
-			</button>
-		{/each}
-		<button
-			aria-label="Annuler"
-			use:tooltip={{ text: 'Annuler', keyboard: '$mod+z', placement: 'right' }}
-			onclick={() => undo.pop()}
-		>
-			<IconUndo />
-		</button>
-		<button
-			aria-label="Rétablir"
-			use:tooltip={{ text: 'Rétablir', keyboard: '$mod+Shift+z', placement: 'right' }}
-			onclick={() => undo.rewind()}
-		>
-			<IconRedo />
-		</button>
+		<Toolbar />
 	</aside>
+
 	<aside class="info">
-		<section class="top">
-			<section class="preactions">
-				<ButtonInk
-					dangerous
-					onclick={deleteImageFileAndGotoNext}
-					help={{
-						text: 'Supprimer cette image et passer à la suivante',
-						keyboard: '$mod+Delete',
-					}}
-				>
-					<IconDelete />
-					Supprimer
-				</ButtonInk>
-
-				<ButtonInk
-					help={{
-						text: "Revenir au recadrage d'origine pour toutes les boîtes",
-
-						keyboard: '$mod+U',
-					}}
-					onclick={revertAll}
-					disabled={!canRevertAll}
-				>
-					<IconRevert />
-					Réinitialiser
-				</ButtonInk>
-			</section>
-		</section>
-		<section class="boxes">
-			<ul>
-				{#each images.filter(({ id }) => id in boundingBoxes) as image, i (image.id)}
-					{@const box = boundingBoxes[image.id]}
-					{@const initBox = initialCrops[image.id]}
-					{@const [w, h] = roundedPixelDimensions(box)}
-					{@const isFocused = focusedImageId === image.id}
-					<li
-						class:unfocused={focusedImageId && focusedImageId !== image.id}
-						class:selected={selectedBox.imageId === image.id}
-						aria-label="Boîte #{i + 1}"
-					>
-						{#if image.fileId && firstImage}
-							<CroppedImg
-								dimensions={firstImage.dimensions}
-								box={toTopLeftCoords(box)}
-								src={uiState.getPreviewURL(image.fileId)}
-								class="thumb"
-							/>
-						{/if}
-						<div class="text">
-							<p class="index">Boîte #{i + 1}</p>
-							<p class="dimensions">
-								<code>
-									<OverflowableText text="{w}×{h}" />
-								</code>
-								<!-- we have a neural-infered value only, put the confidence next to the value -->
-								{#if initBox && !image.metadata[uiState.cropMetadataId].manuallyModified}
-									<span class="sep">&middot;</span>
-									<ConfidencePercentage value={initBox.confidence}>
-										<div class="confidence-icon">
-											<IconNeuralNet />
-										</div>
-									</ConfidencePercentage>
-								{/if}
-							</p>
-						</div>
-						<div class="actions">
-							{#if Object.values(boundingBoxes).length > 1}
-								<ButtonIcon
-									help={isFocused
-										? 'Réafficher les autres boîtes'
-										: 'Masquer les autres boîtes'}
-									keyboard="F"
-									onclick={() => (focusedImageId = isFocused ? '' : image.id)}
-									crossout={isFocused}
-								>
-									<IconFocus />
-								</ButtonIcon>
-							{/if}
-							<ButtonIcon
-								help={initBox
-									? `Revenir au recadrage d'origine 
-									(${roundedPixelDimensions(initBox.value).join(' × ')}, ${percent(initBox.confidence)} de confiance)`
-									: "Recadrage d'origine indisponible"}
-								keyboard="u"
-								disabled={!revertableCrops[image.id]}
-								onclick={() => revertToInferredCrop(image.id)}
-							>
-								<IconRevert />
-							</ButtonIcon>
-							<ButtonIcon
-								help="Supprimer la boîte sélectionnée"
-								keyboard="Delete"
-								onclick={async () => deleteBoundingBox(image.id)}
-							>
-								<IconDelete />
-							</ButtonIcon>
-						</div>
-					</li>
-				{/each}
-				{#if showBoxesListHint}
-					<li class="boxes-list-hint">
-						<p>
-							{@html 'Pour créer une nouvelle boîte,<wbr /> utilisez les outils'}
-							<SentenceJoin items={creationTools} key={(t) => t.name} final="ou">
-								{#snippet children({ icon: Icon, help, shortcut })}
-									<Tooltip text={help} keyboard={shortcut}>
-										<Icon />
-									</Tooltip>
-								{/snippet}
-							</SentenceJoin>
-						</p>
-						<p>
-							Sélectionnez une boîte avec 1 à 9 pour la modifier avec des raccourcis
-							clavier
-						</p>
-					</li>
-				{/if}
-			</ul>
-		</section>
+		<Boxes bind:selectedBox />
 	</aside>
 </div>
 
@@ -1071,214 +279,8 @@
 		overflow: hidden;
 	}
 
-	.toolbar {
-		--width: 2.5em;
-		display: flex;
-		flex-direction: column;
-		align-items: center;
-		padding: 0.25em;
-	}
-
-	.toolbar button {
-		font-size: 1.2em;
-		width: var(--width);
-		height: var(--width);
-		display: flex;
-		justify-content: center;
-		align-items: center;
-		border: none;
-		background: none;
-		cursor: pointer;
-		position: relative;
-		border-radius: var(--corner-radius);
-	}
-
-	.toolbar button.active {
-		color: var(--fg-primary);
-	}
-
-	.toolbar button::after {
-		content: '';
-		position: absolute;
-		bottom: 5px;
-		left: 50%;
-		transform: translateX(-50%);
-		width: 0;
-		height: 3px;
-		border-radius: 1000000px;
-		background: var(--bg-primary);
-		transition: width 0.1s;
-	}
-
-	.toolbar button.active::after {
-		width: 40%;
-	}
-
-	.toolbar button:is(:hover, :focus-visible) {
-		color: var(--fg-primary);
-		background: var(--bg-primary-translucent);
-	}
-
 	.crop-surface {
+		width: 100%;
 		overflow: hidden;
-		position: relative;
-		user-select: none;
-		width: 100%;
-		border-right: 1px solid var(--gray);
-		border-left: 1px solid var(--gray);
-	}
-
-	.behind-image {
-		position: absolute;
-		top: 0;
-		left: 0;
-		width: 100%;
-		height: 100%;
-		display: flex;
-		flex-direction: column;
-		align-items: center;
-		justify-content: center;
-	}
-
-	.behind-image .loading {
-		margin-top: 1em;
-		font-size: 1.2em;
-	}
-
-	.behind-image .coucou {
-		text-align: center;
-		font-size: 0.9em;
-		color: var(--fg-primary);
-	}
-
-	.info {
-		padding: 1em 1.5em;
-		flex: 1;
-		display: flex;
-		flex-direction: column;
-		gap: 1em;
-		min-width: 450px;
-		resize: horizontal;
-		/* to make resize work */
-		overflow: hidden;
-	}
-
-	.info .top {
-		display: flex;
-		flex-direction: column;
-		gap: 0.5em;
-	}
-
-	.info .top .preactions {
-		display: flex;
-		align-items: center;
-		justify-content: space-between;
-		gap: 0.5em;
-	}
-
-	.info h1 {
-		font-size: 1.5em;
-		display: flex;
-		align-items: center;
-		gap: 0.25em;
-	}
-
-	.info h1 .status {
-		display: flex;
-		font-size: 0.7em;
-		color: var(--fg-success);
-	}
-
-	.info section.actions {
-		--width: 100%;
-		display: flex;
-		align-items: center;
-		gap: 0.75em;
-		width: 100%;
-		margin-top: 0.5em;
-	}
-
-	.boxes {
-		height: 100%;
-	}
-
-	.boxes ul {
-		display: flex;
-		flex-direction: column;
-		gap: 0.5em;
-		list-style: none;
-		padding-left: 0;
-	}
-
-	.boxes li {
-		display: flex;
-		gap: 1em;
-	}
-
-	.boxes li:not(.boxes-list-hint) {
-		padding: 0.5em 1em;
-		align-items: center;
-	}
-
-	.boxes li.unfocused {
-		opacity: 0.5;
-	}
-
-	.boxes li.selected {
-		background: var(--bg-primary-translucent);
-	}
-
-	.boxes li.boxes-list-hint {
-		margin-top: 1em;
-		color: var(--gay);
-		flex-direction: column;
-	}
-
-	.boxes li.boxes-list-hint p,
-	.boxes li.boxes-list-hint p :global(.icon) {
-		vertical-align: middle;
-	}
-
-	.boxes li :global(.thumb) {
-		--size: 4rem;
-		width: var(--size);
-		height: var(--size);
-		border-radius: var(--corner-radius);
-		background: color-mix(in srgb, var(--gray) 35%, transparent);
-		flex-shrink: 0;
-	}
-
-	.boxes li .text {
-		overflow: hidden;
-	}
-
-	.boxes li .dimensions {
-		display: flex;
-		align-items: center;
-		gap: 0.5em;
-		font-size: 0.8em;
-
-		&,
-		code {
-			overflow: hidden;
-		}
-	}
-
-	.boxes li .confidence-icon {
-		font-size: 0.8rem;
-		margin-right: 0.5em;
-		display: flex;
-		align-items: center;
-	}
-
-	.boxes li .actions {
-		margin-left: auto;
-		flex-shrink: 0;
-	}
-
-	img {
-		width: 100%;
-		height: 100%;
-		object-fit: contain;
 	}
 </style>
